clients/rl-client-gym/requirements.txt
atari-py==0.0.18
box2d-py==2.3.1
doom-py==0.0.15
gym==0.2.2
imageio==1.5
Keras==1.0.7
mujoco-py==0.5.4
numpy==1.11.1
pachi-py==0.0.21
Pillow==3.3.0
pyglet==1.2.4
PyOpenGL==3.1.0
python-nonblock==3.0.1
PyYAML==3.11
requests==2.10.0
scipy==0.18.0
six==1.10.0
socketIO-client==0.6.6
Theano==0.8.2
websocket-client==0.37.0